[cert] Reorganize the certificates module
I thought a flat structure would work best, but I was wrong.
